[Refactor]#2 presentation.viewmodel 패키지 생성 후 viewmodel 클래스 이동
