+++ app/src/main/java/com/gallery/kakaogallery/presentation/ui/base/BaseFragment.kt
@@ -12,6 +12,7 @@ import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
 import androidx.fragment.app.Fragment
 import com.gallery.kakaogallery.presentation.application.KakaoGalleryApplication
+import com.gallery.kakaogallery.presentation.viewmodel.BaseViewModel
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 
 abstract class BaseFragment<T : ViewDataBinding, R : BaseViewModel> : Fragment() {
