+++ app/src/main/java/com/gallery/kakaogallery/presentation/ui/base/BaseFragmentUseHandler.kt
@@ -3,6 +3,7 @@ package com.gallery.kakaogallery.presentation.ui.base
 import android.content.Context
 import androidx.databinding.ViewDataBinding
 import com.gallery.kakaogallery.presentation.ui.root.FragmentHandler
+import com.gallery.kakaogallery.presentation.viewmodel.BaseViewModel
 
 abstract class BaseFragmentUseHandler<T : ViewDataBinding, R : BaseViewModel> : BaseFragment<T, R>() {
