+++ app/src/main/java/com/gallery/kakaogallery/presentation/ui/gallery/GalleryFragment.kt
@@ -14,6 +14,7 @@ import com.gallery.kakaogallery.R
 import com.gallery.kakaogallery.databinding.FragmentGalleryBinding
 import com.gallery.kakaogallery.presentation.ui.base.BaseFragmentUseHandler
 import com.gallery.kakaogallery.presentation.util.DialogUtil
+import com.gallery.kakaogallery.presentation.viewmodel.GalleryViewModel
 
 class GalleryFragment : BaseFragmentUseHandler<FragmentGalleryBinding, GalleryViewModel>() {
     override val layoutResId: Int
