--- app/src/main/java/com/gallery/kakaogallery/presentation/ui/root/RootViewModel.kt
@@ -1,8 +1,0 @@
-package com.gallery.kakaogallery.presentation.ui.root
-
-import com.gallery.kakaogallery.presentation.ui.base.BaseViewModel
-
-class RootViewModel : BaseViewModel() {
-
-    var currentPage : Int = 0
-}
