+++ app/src/main/java/com/gallery/kakaogallery/presentation/viewmodel/GalleryViewModel.kt
@@ -1,9 +1,8 @@
-package com.gallery.kakaogallery.presentation.ui.gallery
+package com.gallery.kakaogallery.presentation.viewmodel
 
 import android.util.Log
 import com.gallery.kakaogallery.model.ImageModel
 import com.gallery.kakaogallery.model.SaveImageStorage
-import com.gallery.kakaogallery.presentation.ui.base.BaseViewModel
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.subjects.PublishSubject
