+++ app/src/main/java/com/gallery/kakaogallery/presentation/viewmodel/RootViewModel.kt
@@ -1,0 +1,8 @@
+package com.gallery.kakaogallery.presentation.viewmodel
+
+import com.gallery.kakaogallery.presentation.viewmodel.BaseViewModel
+
+class RootViewModel : BaseViewModel() {
+
+    var currentPage : Int = 0
+}
